Move contextprops from src/core to src/context
