+++ test/unit/context/test-contextprops.js
@@ -14,7 +14,11 @@
  * limitations under the License.
  */
 
-import {CanPlay, CanRender, LoadingProp} from '../../../src/core/contextprops';
+import {
+  CanPlay,
+  CanRender,
+  LoadingProp,
+} from '../../../src/context/contextprops';
 
 describes.sandboxed('contextprops - common props', {}, () => {
   describe('CanRender', () => {
